feat: move ClubInfoCard to common, separate organisms from WriteTab
